Change the location of the config file
When this is installed via gems, there won't be a local directory to
save the config file in. This defines the config file as
`~/janrandom.appconfig` so that it should be in a central location. (Is
this is best place for it? Do special permissions need to be defined for
use in multi-user environments? TBD.)