Add some psychographic data
Add some useful psychographic data (this was borrowed from the J-RUG
script by @tomgsmith99) that can be used to flesh out profiles.

Also, separate executable from gem so it can be packaged more easily.
I'm planning on releasing a 0.0.1 gem in a bit when I add methods to
retrieve this interest graph data.